Added timeout for change event and fixed some errors
